--- bindings/python/src/ghex/structured/regular/__init__.py
@@ -1,9 +1,0 @@
-#
-# ghex-org
-#
-# Copyright (c) 2014-2023, ETH Zurich
-# All rights reserved.
-#
-# Please, refer to the LICENSE file in the root directory.
-# SPDX-License-Identifier: BSD-3-Clause
-#
