+++ test/bindings/python/test_structured_pattern.py
@@ -14,11 +14,7 @@ import pytest
 
 from ghex.context import make_context
 from ghex.structured.cartesian_sets import IndexSpace
-from ghex.structured.regular.communication_object import make_communication_object
-from ghex.structured.regular.domain_descriptor import DomainDescriptor
-from ghex.structured.regular.field_descriptor import make_field_descriptor
-from ghex.structured.regular.halo_generator import HaloGenerator
-from ghex.structured.regular.pattern import make_pattern
+from ghex.structured.regular import make_communication_object, DomainDescriptor, make_field_descriptor, HaloGenerator, make_pattern
 
 
 @pytest.mark.mpi
